WC2: Add possibility to allow scenario downgrading later on
In difference to other files, this one cannot be changed later on during 1.16.
Because it is a Lua event, it is loaded on client side.

By defining the variable wc2_highest_player_side behaviour could still be switched during 1.16.

For the code to work, the scenarios must also be loaded.
Currently, only one of the three is loaded, the other scenario ids won't exist.
To load them, the code in WC_II_scenario.cfg would need to be changed on the hosting side to:

#ifdef CAMPAIGN_WC_3P
{WC2_SCENARIO_NEW scenario "WC_II_3p" {WC_II_CAMPAIGN_NAME_3P} 3}
{WC2_SCENARIO_NEW scenario "WC_II_2p" {WC_II_CAMPAIGN_NAME_2P} 2}
{WC2_SCENARIO_NEW scenario "WC_II_1p" {WC_II_CAMPAIGN_NAME_1P} 1}
#endif