WC2: persist color in corner case when using the workaround
This happens if player 4 and another one choose the same color.
(They shouldn't do that to begin with, but they could)

In this case the workaround is also triggered and side 4 gets a new color.
That's fine.
Previously that would have caused side 4's color to change in each scenario,
thus it's now persisted as well.

The workaround is for the case a player chooss choose the same coklor as side 4.