Show the id in the "[race] '' is missing a plural_name" warning
In 1.16 with Ageless Era 4.28.0, the message `[race] '' is missing a
plural_name` is printed to the chat area. The message is not useful even to
add-on developers, because it only identifies which race is affected by
`[race]name`, and in the affected races the `name` attribute is also missing.

An unfortunate side-effect of making the warnings more useful is that more of
them are shown in-game - these messages go through the logging mechanism that
discards duplicate messages. Making them useful also makes them unique.

Also log if a race lacks either a male or female singular name. This warning
will not trigger if the race has a non-empty `name`, as both gender-specific
options are documented to fall back to `name`.

Remove the warning about races that have a `name` but no `id`. Prior to 1.3.8,
[race]name= was the identifier, and there wasn't a [race]id= attribute. The
compatibility code for that change was removed in 1.7.6 (commit 160000a40), so
it seems unnecessary to keep the logging for missing ids.