Fix pango_text::set_maximum_height
Change back to handling the height in our code instead of expecting Pango to
handle it. Pango is still expected to handle the maximum width, as Pango needs
to handle the word-wrapping.

While writing f282eb7, I had noticed that set_maximum_height called
PangoLayout's pango_layout_set_height, and removed the apparently unnecessary
enforcement of the maximum height from our wrapper code. The documentation
added in this commit explains why that was incorrect.

This fixes the bug that some tooltips on the sidebar stopped working. The text
on that sidebar is densely packed, and while this commit makes no visual
change, it removes a few blank pixels of padding below the text. That extra
padding was sufficient for the next line's area to overlap, making
tooltips::add_tooltip remove the previously-added tooltip.

There's still a known issue with 800x600 resolution. At that resolution, the
side-flag is ellipsed (even if the unit has a short name), and the tooltips for
both the name and the flag are missing. The movement and defense tooltips (both
are on the line above the name and flag) are sometimes also missing in 800x600.

(cherry picked from commit 8304ada17cd6db77202b1cd0ffe55a8963980d16)