Chatbox: partially fixed buggy nick autocompletion (fixes #6263)
I don't really like this solution, since as mentioned in #3729, while this works for players
in the lobby, since lobby_info.users isn't updated while a player is in staging/join, anyone
who joins the lobby (or your game) after your create or join a game won't get included in the
autocomplete list. BEeides, I don't like the idea of exposing the full lobby_info object here
just for the user list, but it at least gets ac working and not crashing.