MP/User Info: made state and relation stateless
Instead of storing these as state value which need to be queried, just use query functions.
This allows us to remove the state update code in the lobby and lobby_info, as well as passing
a pointer-to-coint to the Player Info dialog instead of having it manually set the relation state
itself.

This is slightly less efficient when creating the user_info objects since preferences::is_friend
and preferences::is_ignored can now be called twice, but that's negligible. I might return to a
state variable for relation at some point, but in some other guise.