fix cumulative=yes cancel negative value of 'value' in 'leadership'and 'heal'
in 1.16 cumulative=yes don't work in leadership and (i speak of positive value only) the highest value is used, but not added with another, and the use of cumulative=yes cancel negative value beacaude default value is 0.

This is the fix the last issue. For the first, use add or sub instead of 'value'