use "opponent" and "Teacher:" string to specifie origin of abilitie who affect attack
in case of specials abilities with value, the name will be colored after the benefit or not of special.

the mention "Teacher: " and "Opponent:" will be used for what player know what unit attack is affected by Teacher or by opponent.

Remove abilities applied to=opponent is not easy and i want propose that in alternative.