Rework 5e01e13cc85614fd80d4739353142d2c18fd2f0a and e54afdf8b334edfc18750630fb034e4cd539fef4
After closer examination, there was acttually a method to the madness here. Since the "process" functions
may be called more frequently than the "update" functions, we don't want to update the pointer vector and
visibility mask every time the process functions are called. This restores the old order of doing things,
but centralized in lobby_info for clarity.