Minor correction.
The statue Ras-Tabahn refers to is on the 'left' (west) from the player's perspective. However, the party is heading south, and still facing south, when this line is spoken.
Thus, the line really should say to 'our right'. Additionally, there is an animation of the statue being broken straight afterwards so it should be pretty unambiguous as to what is being referred to.

[ci skip]