Avoid counting encounters with 'Cloaked Figure'.
Resolves #6355.

The 'Cloaked Figure' that hounds Kaleh and friends can be encountered up to three times:
* S4 - guaranteed as his appearance is triggered by Kaleh entering the cave which is a scenario requirement.
* S5 - appearance is at a random turn during the battle between the dwarves and the trolls.
* S8 - final appearance is triggered when moving through the narrow passage.

It is possible to miss the encounter in S5 due to the randomised turn mechanic. Thus, it can be confusing when Kaleh says 'third time'.
This is just a quick work-around - another option may be to make the appearance in S5 more deterministic.

[ci skip]