Fix deprecation of [advancefrom], mark as already removed
The existing code searched for `[units][advancefrom]` instead of
`[unit_type][advancefrom]`, so wasn't triggering. It was planned
for removal in 1.17, so simply mark it as already-removed in 1.15.3.

The deprecation warning is shown on stdout/stderr, but not in the
in-game UI, so it's generally not visible to players. I think it
would be better if it was visible, but OTOH it's something that's
easy for someone to check all add-ons for with a simple grep.

(cherry picked from commit f2fa807e6107c63aa9b38f36e5ad8a8fcfacf9a9)