Capitalise direct references to the completed Sceptre of Fire.
Arguably, references to generic sceptres, as is the case with references to the Sceptre of Fire before its completion throughout the rest of this campaign, ought to use the default US English spelling of 'scepter'. This was the case in The Rise of Wesnoth where Haldric quips that the Ruby of Fire should be made into a 'scepter'.

[ci skip]

(cherry picked from commit 101359b06c13c4a5ce310db779125b913f44154e)