Add test scenario for teacher-based weapon specials
This is intended for manual testing, the reason for writing it is to
test tooltips in the attack-predictions dialog. It has a new map, with
five sides packed into adjacent hexes.

Everyone gets a self+allies leadership based on their side number, with
their name in the ability's id and ability's name for tracking whose buffs
are affecting who.