Warning when share_view or share_maps overrides share_vision
Add a new type of deprecation warning, used when a deprecated WML attribute is
used alongside its replacement. Use that to log a deprecation warning if a
[side] has share_vision overridden by share_view or share_maps.

While share_view and share_maps are deprecated on the Wiki, they don't trigger
deprecation warnings on their own. I intend to make them trigger warnings
in 1.17, but they map cleanly to the share_vision enum so the wrapper is easy to
support indefinitely.

There's still one use of the deprecated attributes, in WC's side_definitions.lua.
All SP uses of the deprecated attributes were removed from mainline in
f983a998a2 on the 1.17 branch, which was backported to 1.16 as 80c9263003.