Workaround doxygen issue.
The root cause seems to be a parsing issue in Doxygen, described in #6528. So given this isn't our own code and we want to keep the numbner of patches that need to be applied to a minimum, tell Doxygen to ignore the lua source directory entirely.

This does then (somehow) reveal another issue in another unrelated source file where it's referencing a method in its documentation that doesn't actually exist.

Fixes #6528