Doxygen: fix mainpage links with Doxygen 1.9.1
The extra indentation on these lines caused Doxygen in my local builds
to treat them as a block of preformatted text, and to escape all the HTML.

Locally I'm building with Debian Unstable, which has Doxygen 1.9.1. The
bug isn't present in the CI built docs on https://devdocs.wesnoth.org/
which were built with Doxygen 1.8.13.