SoF: explain the Shorbear tools as products of the volcanic forge
I don't like the explanation of the tools being made of unobtainium, so changed
it to use a convenient plot device. I think this also solves a plot hole about
the forge; instead of it being near the Shorbear Caves by random chance, this
gives a reason for the Shorbear Clan to live there (even if they haven't been
in there in living memory).