Convert the tutorial's wall of text [message]s into overlay text
Adds the implementation of a prototype [hint_message] tag, which will either use
PR 5837's add_overlay_text, or a fallback of outputting them to the chat log.
This is partly to review PR 5873 itself, seeing how it works for this UX.

This means a few less dialogs for the user to click through, but mainly means
that about one hint per turn stays visible even when skipping through the
dialogs.

Turn 1's hint about upkeep appears near the end, when the player captures a village.

Turn 2's hint appears in the start-of-turn dialogue, when commenting on the orc
blocking the bridge.

Turn 3 onwards appear at the start of each turn.

Initially I tried the bottom of the screen, but that overlaps with the area for
`[message]`s.