Tutorial S02: Reduce the number of [messages]
Drop many of the flavortext messages, to focus on the ones that explain
mechanics or tactics to the player.

The message about the Status Table gets removed because the player is unlikely
to benefit from it in this scenario. It might fit near the end as "you can
check how much gold the enemy has, and guess if he'll recruit next turn", but
otherwise it's just "hey, you could look at all the menu items" without any
suggestion about why the player should use it now.