Add debug for "[event] is missing name or id"
The problem with this warning was that it didn't include any clue about what
had triggered it - just that there's a broken event somewhere, and it could be
in a scenario, macro, unit_type or even from another add-on if you're using a
non-default era.

When run with `--log-debug=wml`, this makes Wesnoth print out whatever the
contents of the affected [event] tag are.

(cherry picked from commit d6b56969a012853fbfc55844b1eac98a70c0346b)