Update pixel scale in real-time from prefs dialog.
This will never reduce the drawing resolution beyond the minimum window
size specified in preferences/general.cpp.

The minimum window height has been reduced from 600px to 540px!
This is to that a 1920x1080 display can use a pixel scale of 2.
It could be changed back, but a better solutuion would be updating the
default theme to accomodate this form factor.