Partial pot-update: trivial changes only
These are typo-fixes and minor grammar improvements to the English text,
which don't change the meaning of the text and therefore shouldn't set
the fuzzy flags on their respective translations.

This is a hand-edited subset of the changes that a full pot-update would make.
The plan once this is reviewed is to update the .po files with the changes in
these .pot files, but to filter out any added fuzzy flags when commiting them
to Git. A normal pot-update can then be run, which will add fuzzy flags for the
string changes which weren't included in this PR.