pofix: wesnoth-sof rules for 1.16.3
This will still leave 20 fuzzies that are trivial but which should
not be done with a line-by-line search-and-replace.

The string beginning "Much to his chagrin, Alanin was dispatched to scout
further" has multiple changes, some of which split over the word-wrap. Two
parts of it get fixed here, but it will still become fuzzy due to the changes
around "cold passage" part.